This boots the DB and passes a test.
