+++ imls-backend/init/010-schemas.sql
@@ -1,5 +1,3 @@
--- migrate:up
-
 --
 -- Name: admin; Type: SCHEMA; Schema: -; Owner: -
 --
@@ -10,9 +8,7 @@ CREATE SCHEMA admin;
 --
 -- Name: api; Type: SCHEMA; Schema: -; Owner: -
 --
-
 CREATE SCHEMA api;
-
 
 --
 -- Name: data; Type: SCHEMA; Schema: -; Owner: -
@@ -20,11 +16,8 @@ CREATE SCHEMA api;
 
 CREATE SCHEMA data;
 
-
 --
 -- Name: imlswifi; Type: SCHEMA; Schema: -; Owner: -
 --
 
 CREATE SCHEMA imlswifi;
-
--- migrate:down
