+++ imls-backend/init/011-roles.sql
@@ -1,0 +1,2 @@
+
+CREATE ROLE web_anon NOLOGIN;
