+++ imls-backend/init/100-config.sql
@@ -1,5 +1,3 @@
--- migrate:up
-
 SET statement_timeout = 0;
 SET lock_timeout = 0;
 SET idle_in_transaction_session_timeout = 0;
@@ -10,5 +8,3 @@ SET check_function_bodies = false;
 SET xmloption = content;
 SET client_min_messages = warning;
 SET row_security = off;
-
--- migrate:down
