+++ imls-backend/init/130-views.sql
@@ -12,12 +12,13 @@ CREATE VIEW admin.libraries AS
 --
 -- Name: imls_lookup; Type: VIEW; Schema: api; Owner: -
 --
-
-CREATE VIEW api.imls_lookup AS
- SELECT imls_lookup.id,
-    imls_lookup.fscs_id,
-    imls_lookup.timezone
-   FROM imlswifi.imls_lookup;
+-- NOTE: What is the difference, really, between these two?
+CREATE VIEW api.imls_lookup AS SELECT * FROM imlswifi.imls_lookup;
+-- CREATE VIEW api.imls_lookup AS
+--  SELECT imls_lookup.id,
+--     imls_lookup.fscs_id,
+--     imls_lookup.timezone
+--    FROM imlswifi.imls_lookup;
 
 
 --
