+++ imls-backend/init/400-permissions.sql
@@ -1,0 +1,5 @@
+
+GRANT USAGE ON SCHEMA api TO web_anon;
+GRANT USAGE ON SCHEMA data TO web_anon;
+
+GRANT SELECT ON TABLE api.imls_lookup TO web_anon;
